polaroid-python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel", "grpcio-tools>=1.60.0"]
build-backend = "setuptools.build_meta"

[project]
name = "polaroid"
version = "0.1.0"
description = "FDAP-optimized DataFrame engine with gRPC interface"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Polaroid Contributors"}
]
keywords = ["dataframe", "arrow", "grpc", "streaming", "time-series"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "grpcio>=1.60.0",
    "grpcio-tools>=1.60.0",
    "pyarrow>=15.0.0",
    "protobuf>=4.25.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.urls]
Homepage = "https://github.com/EnkiNudimmud/polaroid"
Documentation = "https://docs.polaroid.rs"
Repository = "https://github.com/EnkiNudimmud/polaroid"
Issues = "https://github.com/EnkiNudimmud/polaroid/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["polaroid*"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
